requirements.txt
numpy==2.3.3
scipy==1.16.2
matplotlib==3.10.7
pandas==2.3.3
PyYAML==6.0.3
tqdm==4.67.1
